requirements.txt
faster-whisper
sentencepiece
torch
transformers
pyannote.audio
python-dotenv
soundfile
